Move validate_dataset to the config.py file
