+++ kpconv_torch/cli/__main__.py
@@ -5,17 +5,6 @@ from kpconv_torch import __version__ as kpconv_version
 from kpconv_torch import preprocess, test, train
 
 
-SUPPORTED_DATASETS = {"ModelNet40", "NPM3D", "S3DIS", "SemanticKitti", "Toronto3D"}
-
-
-def valid_dataset(dataset):
-    if dataset not in SUPPORTED_DATASETS:
-        raise argparse.ArgumentTypeError(
-            f"{dataset} dataset is unknown, please choose amongst {SUPPORTED_DATASETS}."
-        )
-    return dataset
-
-
 def valid_dir(str_dir):
     """Build a ``pathlib.Path`` object starting from the ``str_dir`` folder."""
     path_dir = Path(str_dir)
